+++ chatroom/templates/chatroom/chatroom.html
@@ -10,7 +10,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Global Chat Room</title>
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="static/chatroom/css/style.css">
+    <link rel="stylesheet" href="{% static 'chatroom/css/style.css' %}">
 </head>
 <body>
     
@@ -119,7 +119,7 @@
             });
         });
     </script>
-    <script src="{% static 'chatroom/js/endScript.js' %}"></script>
+<script src="{% static 'chatroom/js/startScript.js' %}"></script>
 
     
 </body>
